moves new executor to libs and renames it to execute
