+++ openvair/modules/tools/utils.py
@@ -63,7 +63,6 @@ from sqlalchemy.exc import OperationalError
 
 from openvair import config
 from openvair.libs.log import get_logger
-from openvair.models.execution_result import ExecutionResult
 
 if TYPE_CHECKING:
     from sqlalchemy.orm import Session
@@ -238,175 +237,6 @@ def execute(*cmd: str, **kwargs: Any) -> Tuple[str, str]:  # noqa: C901 ANN401 n
     except OSError as err:
         message = 'Got an OSError\ncommand: %(cmd)r\nerrno: %(errno)r'
         LOG.info(message, {'cmd': cmd, 'errno': err.errno})
-        raise
-
-
-def terminate_process(proc: subprocess.Popen, cmd_str: str) -> str:
-    """Attempts to gracefully terminate a subprocess and force kill it if fails.
-
-    Args:
-        proc (subprocess.Popen): The subprocess to be terminated.
-        cmd_str (str): The command string representing the subprocess for
-            logging purposes.
-
-    Returns:
-        str: The standard error output (stderr) collected from the subprocess
-            after termination.
-    """
-    LOG.warning(f"Command '{cmd_str}' timed out. Terminating process.")
-    proc.terminate()
-    try:
-        proc.wait(timeout=5)
-    except subprocess.TimeoutExpired:
-        LOG.error(f"Command '{cmd_str}' did not terminate. Killing it.")
-        proc.kill()
-    _, stderr = proc.communicate()
-    return str(stderr)
-
-
-def __run_command(
-    proc: subprocess.Popen,
-    cmd_str: str,
-    timeout: Optional[float],
-    *,
-    raise_on_error: bool,
-) -> ExecutionResult:
-    """Handles the execution of a subprocess and processes its results.
-
-    Args:
-        proc (subprocess.Popen): The subprocess to be executed.
-        cmd_str (str): The string representation of the command being executed.
-        timeout (Optional[float]): Maximum time in seconds to wait for the
-            command to complete. Defaults to None (no timeout).
-        raise_on_error (bool): If True, raises an exception if the command exits
-            with a non-zero return code. If False, includes the return code
-            in the output for manual handling.
-
-    Returns:
-        ExecutionResult: An object containing the following fields:
-            - returncode (int): The exit code of the command.
-            - stdout (str): Standard output of the command.
-            - stderr (str): Standard error output of the command.
-
-    Raises:
-        ExecuteTimeoutExpiredError: Raised if the command execution exceeds the
-            specified timeout.
-        ExecuteError: Raised if the command exits with a non-zero return code
-            and raise_on_error is True.
-    """
-    try:
-        stdout, stderr = proc.communicate(timeout=timeout)
-        returncode = proc.returncode
-        LOG.info(f"Command '{cmd_str}' completed with return code {returncode}")
-        if raise_on_error and returncode != 0:
-            message = (
-                f"Command '{cmd_str}' failed with return code {returncode}"
-            )
-            LOG.error(message)
-            raise ExecuteError(message)
-
-        return ExecutionResult(
-            returncode=returncode,
-            stdout=stdout or '',
-            stderr=stderr or '',
-        )
-
-    except subprocess.TimeoutExpired:
-        stderr = terminate_process(proc, cmd_str)
-
-        LOG.warning(
-            f"Command '{cmd_str}' timed out. Attempting graceful termination."
-        )
-        proc.terminate()  # soft terminate
-        try:
-            proc.wait(timeout=5)  # 5 seconds for graceful termination
-        except subprocess.TimeoutExpired:
-            LOG.error(f"Command '{cmd_str}' did not terminate. Killing it now.")
-            proc.kill()  # kill process after 5 seconds if not terminated
-        _, stderr = proc.communicate()  # Get stderr after finishing process
-        message = (
-            f"Command '{cmd_str}' timed out and was killed.\n"
-            f'Error: {stderr.decode()}'
-        )
-        raise ExecuteTimeoutExpiredError(message)
-
-
-def execute_command(
-    *args: str,
-    shell: bool = False,
-    run_as_root: bool = False,
-    root_helper: str = 'sudo',
-    timeout: Optional[float] = None,
-    raise_on_error: bool = False,
-) -> ExecutionResult:
-    """Executes a shell command and returns its stdout, stderr, and exit code.
-
-    Args:
-        *args (str): The command and its arguments to execute. Each argument
-            must be passed as a separate string.
-        shell (bool): If True, the command will be executed through the shell.
-            Defaults to False.
-        run_as_root (bool): If True, the command will be executed with root
-            privileges using the specified root_helper. Defaults to False.
-        root_helper (str): The command used to elevate privileges, such as
-            'sudo'. Defaults to 'sudo'.
-        timeout (Optional[float]): Maximum time in seconds to wait for the
-            command to complete. Defaults to None (no timeout).
-        raise_on_error (bool): If True, raises an exception if the command exits
-            with a non-zero return code. If False, includes the return code
-            in the output dictionary for manual handling. Defaults to False.
-
-    Returns:
-        Dict[str, Union[str, int]]: A dictionary containing the following keys:
-            - 'stdout': Standard output of the command (str).
-            - 'stderr': Standard error output of the command (str).
-            - 'returncode': The exit code of the command (int). This is included
-                only if raise_on_error is False.
-
-    Raises:
-        ExecuteTimeoutExpiredError: Raised if the command execution exceeds the
-            specified timeout.
-        ExecuteError: Raised if the command exits with a non-zero return code
-            and raise_on_error is True.
-        OSError: Raised for system-level errors, such as command not found or
-            permission issues.
-
-    Example:
-        >>> execute2('ls', '-l', raise_on_error=True)
-        {'stdout': 'output', 'stderr': '', 'returncode': 0}
-
-        >>> execute2('false', raise_on_error=True)
-        Traceback (most recent call last):
-            ...
-        ExecuteError: Command 'false' failed
-
-        >>> result = execute2('false', raise_on_error=False)
-        >>> print(result)
-        {'stdout': '', 'stderr': '', 'returncode': 1}
-    """
-    cmd: List[str] = list(args)
-    if run_as_root and hasattr(os, 'geteuid') and os.geteuid() != 0:
-        cmd = [root_helper, *cmd]
-
-    cmd_str = ' '.join(cmd)
-    LOG.info(f'Executing command: {cmd_str}')
-    try:
-        with subprocess.Popen(  # noqa: S603
-            cmd_str if shell else cmd,
-            shell=shell,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            stdin=subprocess.PIPE,
-            text=True,
-        ) as proc:
-            return __run_command(
-                proc,
-                cmd_str,
-                timeout,
-                raise_on_error=raise_on_error,
-            )
-    except OSError as err:
-        LOG.error(f"OS error for command '{cmd_str}': {err}")
         raise
